requirements.txt
flask>=3.0
flask-cors>=4.0
yfinance>=0.2
finnhub-python>=2.4
pandas>=2.0,<3.0
matplotlib>=3.8
requests>=2.31
python-dotenv>=1.0
financedatabase>=2.3
financetoolkit>=2.0.3,<3.0
openpyxl>=3.1
scikit-learn>=1.6
tqdm>=4.67
